+++ HISTORY.rst
@@ -144,6 +144,7 @@
   that document type is removed from the workflow.
 - Make error messages persistent and increase the timeout of warning to 10 seconds.
 - Improve rendering of the details form.
+- Update rendering of the readonly multiselect widget to conform to Django's updated field class interface.
 
 2.7.3 (2017-09-11)
 ==================